backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pandas==2.2.3
openpyxl==3.1.5
python-multipart==0.0.19
supabase==2.11.0
jinja2==3.1.5
requests==2.32.3
python-dotenv==1.0.1
pydantic-settings==2.7.1
pyjwt==2.10.1
httpx==0.28.1
weasyprint==63.1
